move types out of app-render
