+++ packages/next/src/client/components/router-reducer/apply-router-state-patch-to-tree.ts
@@ -1,7 +1,7 @@
 import type {
   FlightRouterState,
   FlightSegmentPath,
-} from '../../../server/app-render'
+} from '../../../server/app-render/types'
 import { matchSegment } from '../match-segments'
 
 /**
