+++ packages/next/src/client/components/router-reducer/fill-lazy-items-till-leaf-with-head.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { fillLazyItemsTillLeafWithHead } from './fill-lazy-items-till-leaf-with-head'
 import { CacheStates, CacheNode } from '../../../shared/lib/app-router-context'
-import { FlightData } from '../../../server/app-render'
+import type { FlightData } from '../../../server/app-render/types'
 
 const getFlightData = (): FlightData => {
   return [
