+++ packages/next/src/client/components/router-reducer/invalidate-cache-below-flight-segmentpath.ts
@@ -1,5 +1,5 @@
 import type { CacheNode } from '../../../shared/lib/app-router-context'
-import type { FlightSegmentPath } from '../../../server/app-render'
+import type { FlightSegmentPath } from '../../../server/app-render/types'
 
 /**
  * Fill cache up to the end of the flightSegmentPath, invalidating anything below it.
