+++ packages/next/src/client/components/router-reducer/invalidate-cache-by-router-state.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { invalidateCacheByRouterState } from './invalidate-cache-by-router-state'
 import { CacheStates, CacheNode } from '../../../shared/lib/app-router-context'
-import { FlightRouterState } from '../../../server/app-render'
+import type { FlightRouterState } from '../../../server/app-render/types'
 
 describe('invalidateCacheByRouterState', () => {
   it('should invalidate the cache by router state', () => {
