+++ packages/next/src/client/components/router-reducer/is-navigating-to-new-root-layout.ts
@@ -1,4 +1,4 @@
-import type { FlightRouterState } from '../../../server/app-render'
+import type { FlightRouterState } from '../../../server/app-render/types'
 
 export function isNavigatingToNewRootLayout(
   currentTree: FlightRouterState,
